Fix $cassandra_seed value now that we only listen on eth0.
Also remove seed setup for a basic host from Hiera.  We're relying on
eth0 for addersses and I'd rather not set some values in the manifest
and others in Hiera.  We'll make sure a single host can get up and work
without error.  We'll work on handling actual clusters via
Hiera/ENC/Vagrant.